chore(server): refactor large file kubernetes_service.py
