+++ server/opensandbox_server/services/k8s/client.py
@@ -31,7 +31,6 @@ from opensandbox_server.services.k8s.rate_limiter import TokenBucketRateLimiter
 
 logger = logging.getLogger(__name__)
 
-# Type alias for informer cache key
 _InformerKey = Tuple[str, str, str, str]  # (group, version, plural, namespace)
 
 
@@ -49,10 +48,8 @@ class K8sClient:
         self._core_v1_api: Optional[CoreV1Api] = None
         self._custom_objects_api: Optional[CustomObjectsApi] = None
         self._node_v1_api: Optional[NodeV1Api] = None
-        # Informer pool: key -> WorkloadInformer
         self._informers: Dict[_InformerKey, WorkloadInformer] = {}
         self._informers_lock = threading.Lock()
-        # Rate limiters (None = unlimited)
         self._read_limiter: Optional[TokenBucketRateLimiter] = (
             TokenBucketRateLimiter(qps=k8s_config.read_qps, burst=k8s_config.read_burst)
             if k8s_config.read_qps > 0
@@ -63,10 +60,6 @@ class K8sClient:
             if k8s_config.write_qps > 0
             else None
         )
-
-    # ------------------------------------------------------------------
-    # Internal API handle accessors (lazy singletons)
-    # ------------------------------------------------------------------
 
     def _load_config(self) -> None:
         """Load kubeconfig from file path or in-cluster service account."""
@@ -93,9 +86,6 @@ class K8sClient:
             self._node_v1_api = client.NodeV1Api()
         return self._node_v1_api
 
-    # ------------------------------------------------------------------
-    # Internal informer pool management
-    # ------------------------------------------------------------------
 
     def _get_informer(self, group: str, version: str, plural: str, namespace: str) -> Optional[WorkloadInformer]:
         """Return the informer for this resource+namespace, starting it lazily."""
@@ -128,9 +118,6 @@ class K8sClient:
                     return None
         return informer
 
-    # ------------------------------------------------------------------
-    # CustomObject operations
-    # ------------------------------------------------------------------
 
     def create_custom_object(
         self,
@@ -255,9 +242,6 @@ class K8sClient:
             body=body,
         )
 
-    # ------------------------------------------------------------------
-    # Secret operations
-    # ------------------------------------------------------------------
 
     def create_secret(self, namespace: str, body: Any) -> Any:
         """Create a namespaced Secret."""
@@ -268,9 +252,6 @@ class K8sClient:
             body=body,
         )
 
-    # ------------------------------------------------------------------
-    # Pod operations
-    # ------------------------------------------------------------------
 
     def list_pods(
         self,
@@ -285,10 +266,6 @@ class K8sClient:
             label_selector=label_selector,
         )
         return resp.items
-
-    # ------------------------------------------------------------------
-    # RuntimeClass operations
-    # ------------------------------------------------------------------
 
     def read_runtime_class(self, name: str) -> Any:
         """Read a RuntimeClass from the cluster."""
