+++ server/opensandbox_server/services/k8s/status_helpers.py
@@ -12,4 +12,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Packaged example configuration templates for the OpenSandbox server."""
+from __future__ import annotations
+
+from typing import Any, Dict
+
+
+def _normalize_create_status(status_info: Dict[str, Any]) -> Dict[str, Any]:
+    if status_info.get("state") != "Allocated":
+        return status_info
+    return {
+        **status_info,
+        "state": "Running",
+        "message": "Pod has IP assigned and sandbox is ready for requests",
+    }
+
+
+def _is_unschedulable_status(status_info: Dict[str, Any]) -> bool:
+    reason = str(status_info.get("reason") or "")
+    return reason == "POD_PLATFORM_UNSCHEDULABLE"
