+++ server/opensandbox_server/services/k8s/volume_helper.py
@@ -28,23 +28,7 @@ def apply_volumes_to_pod_spec(
     pod_spec: Dict[str, Any],
     volumes: List[Volume],
 ) -> None:
-    """
-    Apply user-specified volumes to a Kubernetes pod spec.
-
-    This function converts Volume API objects to Kubernetes volume and volumeMount
-    definitions and adds them to the pod spec in-place.
-
-    Currently supported backends:
-    - pvc: Maps to Kubernetes PersistentVolumeClaim
-    - host: Maps to Kubernetes hostPath volume
-
-    Args:
-        pod_spec: The pod spec dictionary to modify in-place
-        volumes: List of Volume API objects
-
-    Raises:
-        ValueError: If an unsupported volume backend is specified
-    """
+    """Apply user-specified volumes to a pod spec in-place."""
     containers = pod_spec.get("containers", [])
     if not containers:
         logger.warning("No containers in pod spec, skipping volume mounts")
@@ -54,15 +38,12 @@ def apply_volumes_to_pod_spec(
     mounts = main_container.get("volumeMounts", [])
     pod_volumes = pod_spec.get("volumes", [])
 
-    # Collect existing volume names to prevent collisions with internal volumes
     existing_volume_names = {v.get("name") for v in pod_volumes if isinstance(v, dict)}
-    # One Kubernetes volume per unique PVC; multiple volumeMounts can reference it
     pvc_to_volume_name: Dict[str, str] = {}
 
     for vol in volumes:
         vol_name = vol.name
 
-        # Check for collision with internal volumes
         if vol_name in existing_volume_names:
             raise ValueError(
                 f"Volume name '{vol_name}' conflicts with an internal volume. "
@@ -70,14 +51,9 @@ def apply_volumes_to_pod_spec(
             )
 
         if vol.pvc is not None:
-            # PVC backend: maps to Kubernetes PersistentVolumeClaim.
-            # Multiple Volume API objects sharing the same claim_name must produce
-            # a single Kubernetes volume and multiple volumeMounts (CSI drivers
-            # can fail when the same PVC is defined in multiple volume entries).
             pvc_claim_name = vol.pvc.claim_name
 
             if pvc_claim_name not in pvc_to_volume_name:
-                # First use of this PVC: create one volume, use current vol.name as volume name
                 pod_volumes.append({
                     "name": vol_name,
                     "persistentVolumeClaim": {
@@ -100,8 +76,6 @@ def apply_volumes_to_pod_spec(
                 f"Added PVC volume '{vol_name}' (claim: {pvc_claim_name}) mounted at '{vol.mount_path}' for sandbox"
             )
         elif vol.host is not None:
-            # Host backend: maps to hostPath volume
-            # Note: hostPath is node-local and not recommended for production
             host_path = vol.host.path
 
             pod_volumes.append({
@@ -130,6 +104,5 @@ def apply_volumes_to_pod_spec(
                 "Supported backends: pvc, host"
             )
 
-    # Update pod spec with modified volumes and mounts
     pod_spec["volumes"] = pod_volumes
     main_container["volumeMounts"] = mounts
